refactor: 💡 separate out utils
