+++ src/components/DataRoom/USDCStorage/utils/addWallsToWorld.ts
@@ -1,16 +1,9 @@
 import Matter from 'matter-js'
 const { Bodies, World } = Matter
+import type { Dimensions } from './types'
 
 const WALL_BORDER_WIDTH = 25
 const WALL_LENGTH = 500
-const IMG_TEXTURE_SIZE = 256 // Size Of The USDC.png Image In Pixels
-const COIN_RADIUS_SM = 35
-const COIN_RADIUS_MD = 50
-
-export type Dimensions = {
-  width: number
-  height: number
-}
 
 export const createWalls = (dimensions: Dimensions) => [
   { x: 0, y: -WALL_BORDER_WIDTH / 2, width: WALL_LENGTH * 8, height: WALL_BORDER_WIDTH },
@@ -34,20 +27,4 @@ export const addWallsToWorld = (engine: Matter.Engine, walls: any[]) => {
       }),
     ),
   )
-}
-
-export const createCoin = (dimensions: Dimensions, imgUrl: string, isMobile: boolean) => {
-  const coinRadius = isMobile ? COIN_RADIUS_SM : COIN_RADIUS_MD
-  const IMG_SCALE = (coinRadius * 2) / IMG_TEXTURE_SIZE // Scale Factor To Match Image Texture Size With Coin Radius
-
-  return Bodies.circle(Math.random() * dimensions.width * 0.5 + dimensions.width * 0.25, 0, coinRadius, {
-    restitution: 0.4,
-    render: {
-      sprite: {
-        texture: imgUrl,
-        xScale: IMG_SCALE,
-        yScale: IMG_SCALE,
-      },
-    },
-  })
 }
