+++ src/components/DataRoom/USDCStorage/utils/types.ts
@@ -1,0 +1,4 @@
+export type Dimensions = {
+  width: number
+  height: number
+}
